Actualización en la aplicación. La información lee tarjetas ahora de un Arduino y guarda información de tarjetas RFID

+++ Views/Informacion_empleado.cs
@@ -9,6 +9,7 @@ using System.Linq;
 using System.Numerics;
 using System.Runtime.CompilerServices;
 using System.Text;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
@@ -24,7 +25,7 @@ namespace PAS.Views
             Genero.Items.Add("Femenino");
             Genero.Items.Add("Otro");
             this.inicio = inicio;
-
+            this.inicio.Visible = false;
         }
 
 
@@ -62,26 +63,88 @@ namespace PAS.Views
         {
             int opcion = this.acciones.SelectedIndex;
             Usuario usuario = new Usuario();
+            string sql = "";
+
+
+            // Filtros
+            if (!(opcion == 1 || opcion == 3))
+            {
+                if (!isPhoneNumber(telefono.Text))
+                {
+                    MessageBox.Show("Ingrese bien el número de teléfono.");
+                    return;
+                }
+
+                if (!validContract(inicio_contrato, final))
+                {
+                    MessageBox.Show("Tiene que ingresar un rango de contrato válido");
+                    return;
+                }
+            }
+
             switch (opcion)
             {
                 case 0:
                     this.llenarDatos(usuario);
-                    Queries.sendToDataBase(usuario.AddToDataBase());
+
+                    // Filtro de registro.
+                    string isRegistered = usuario.searchUser(usuario.Identificacion);
+                    string isCardRegistered = usuario.searchByCard(usuario.Tarjeta);
+
+                    if (Queries.results(isRegistered).Rows.Count > 0) {
+                        MessageBox.Show("Persona registrada con anterioridad.");
+                        return;
+                    }
+
+                    if (cardResponse.Text.Length <= 0) {
+                        MessageBox.Show("Tarjeta no leída.");
+                        return;
+                    }
+
+                    if (Queries.results(isCardRegistered).Rows.Count > 0) {
+                        MessageBox.Show("Tarjeta agregada con anterioridad.");
+                        return;
+                    }
+
+                    sql = usuario.AddToDataBase();
+                    MessageBox.Show(sql);
+                    Queries.sendToDataBase(sql);
+
+                    sql = ListOfResults.crearContrato("" + usuario.Identificacion, usuario.Inicio, usuario.Final);
+                    Queries.sendToDataBase(sql);
+                    
+                    this.limpiarCampos();
                     break;
+
                 case 1:
-                    DataTable table = Queries.results(usuario.searchUser(BigInteger.Parse(this.identificacion.Text)));
-                    if (table == null) { MessageBox.Show("Empleado no encontrado."); break; }
+                    DataTable table = null;
+                    if (identificacion.Text.Length > 0)
+                        table = Queries.results(usuario.searchUser(BigInteger.Parse(this.identificacion.Text)));
+                    else    
+                        table = Queries.results(usuario.searchByCard(this.cardResponse.Text));
+                    
+
+                    if (table.Rows.Count <= 0) { MessageBox.Show("Empleado no encontrado."); break; }
+
                     usuario.toClass(table.Rows[0]);
                     llenarCampos(usuario);
                     break;
+
                 case 2:
                     this.llenarDatos(usuario);
                     Queries.sendToDataBase(usuario.updateUser());
+
+                    sql = ListOfResults.actualizarContrato("" + usuario.Identificacion, usuario.Inicio, usuario.Final);
+
+                    Queries.sendToDataBase(sql);
+                    this.limpiarCampos();
                     break;
                 case 3:
                     DialogResult result = MessageBox.Show("¿Quieres realizar esta acción?", "Confirmación", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-                    if (result == DialogResult.Yes) {
+                    if (result == DialogResult.Yes)
+                    {
                         Queries.sendToDataBase(usuario.deleteUser(BigInteger.Parse(this.identificacion.Text)));
+                        this.limpiarCampos();
                     }
                     break;
                 default: MessageBox.Show("Ingrese una opción válida."); break;
@@ -101,7 +164,6 @@ namespace PAS.Views
             this.telefono.Enabled = condicion;
             this.direccion.Enabled = condicion;
             this.Genero.Enabled = condicion;
-            this.huella.Enabled = condicion;
             this.inicio_contrato.Enabled = condicion;
             this.final.Enabled = condicion;
             this.correo.Enabled = condicion;
@@ -116,12 +178,15 @@ namespace PAS.Views
             usuario.setTelefono(this.telefono.Text);
             usuario.setGenero((string)this.Genero.SelectedItem);
             usuario.setDireccion(this.direccion.Text);
-            usuario.setInicio(this.inicio_contrato.Value.ToShortDateString());
-            usuario.setFinal(this.final.Value.ToShortDateString());
+            usuario.setInicio(this.inicio_contrato.Value.ToString("yyyy-MM-dd"));
+            usuario.setFinal(this.final.Value.ToString("yyyy-MM-dd"));
             usuario.setCorreo(this.correo.Text);
+            usuario.setTarjeta(this.cardResponse.Text);
         }
 
-        private void llenarCampos(Usuario usuario) {
+        private void llenarCampos(Usuario usuario)
+        {
+            this.identificacion.Text = "" + usuario.Identificacion;
             this.nombre.Text = usuario.Nombre;
             this.apellido.Text = usuario.Apellido;
             this.telefono.Text = usuario.Telefono;
@@ -131,16 +196,44 @@ namespace PAS.Views
             if (genero.Equals("Otro")) { this.Genero.SelectedIndex = 2; }
             this.direccion.Text = usuario.Direccion;
             this.correo.Text = usuario.Correo;
+            this.cardResponse.Text = usuario.Tarjeta;
         }
 
         private void limpiarCampos()
         {
+            this.identificacion.Text = "";
             this.nombre.Text = "";
             this.apellido.Text = "";
             this.telefono.Text = "";
             this.Genero.SelectedIndex = 0;
             this.direccion.Text = "";
             this.correo.Text = "";
+        }
+
+        private Boolean isPhoneNumber(string telefono)
+        {
+            string pattern = @"^\d{10}$";
+            return Regex.IsMatch(telefono, pattern);
+        }
+
+        private Boolean validContract(DateTimePicker inicio, DateTimePicker fin)
+        {
+            DateTime dateInicio = inicio.Value, dateFinal = fin.Value;
+            MessageBox.Show(dateInicio + " " + dateFinal + " = " + DateTime.Compare(dateInicio, dateFinal));
+            return DateTime.Compare(dateInicio, dateFinal) <= 0;
+        }
+
+
+        private void huella_Click(object sender, EventArgs e)
+        {
+            string data = Arduino.ArduinoTarjeta.SeriaConnection();
+            cardResponse.Text = data;
+
+        }
+
+        private void button2_Click(object sender, EventArgs e)
+        {
+            MessageBox.Show("Ingrese el número de teléfono sin espacios.");
         }
     }
 }
